Rename GetDevice and GetState commands/arguments to Device and State
